рефакторинг ver 2
не создает задачу

+++ app/Http/Controllers/Task/DeleteController.php
@@ -2,16 +2,35 @@
 
 namespace App\Http\Controllers\Task;
 
-use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Models\Task;
+use App\Client\MoySkladClient;
+use App\Http\Controllers\HomeController;
+use Illuminate\Http\Request;
 
-class DeleteController extends Controller
-{
-    public function __invoke(Task $task)
+class DeleteController extends HomeController {
+
+    public function __construct(MoySkladClient $msClient)
     {
-        $task->delete();
-        return response([]);
+        $this->msClient = $msClient;
+    }
+
+    public function __invoke($id)
+    {
+        $task = Task::find($id);
+
+        if (!$task) {
+            return response()->json(['error' => 'Task not found'], 404);
+        }
+
+        $ms_uuid = $task->ms_uuid;
+        $deleted = $this->msClient->deleteTask($ms_uuid);
+
+        if ($deleted) {
+            $task->delete();
+            return response()->json(null, 204);
+        }
+
+        return response()->json(['error' => 'Failed to delete task in MoySklad'], 500);
     }
 }
-
